Passed back to f32 for now for the sim.

+++ examples/cu_rp_balancebot/src/sim.rs
@@ -147,7 +147,7 @@ fn run_copper_callback(
                     }
                     let force_magnitude = motor_actuation.power * 2.0; // 4_000.0;
                                                                        // let current_force = cart_force.force(); // Get existing force which includes gravity
-                    let new_force = /*current_force */ Vector::new(force_magnitude as f64, 0.0, 0.0);
+                    let new_force = /*current_force */ Vector::new(force_magnitude, 0.0, 0.0);
                     cart_force.apply_force(new_force);
                     output
                         .metadata
